[TASK-1] Fix8

+++ boris.martynov/task-1/internal/calculator/calculator.go
@@ -1,57 +1,57 @@
 package calc
 
 import (
- "fmt"
+	"fmt"
 )
 
 func Add(a, b int64) int64 {
- return a + b
+	return a + b
 }
 
 func Subt(a, b int64) int64 {
- return a - b
+	return a - b
 }
 
 func Mul(a, b int64) int64 {
- return a * b
+	return a * b
 }
 
-func Div(a, b int64) float64 {
- if b == 0 {
-  fmt.Println("Division by zero")
-  return 0
- }
- return float64(a) / float64(b)
+func Div(a, b int64) int64 {
+	if b == 0 {
+		fmt.Println("Division by zero")
+		return 0
+	}
+	return a / b
 }
 
 func UserInput() (int64, int64, string) {
- var a, b int64
- var operator string
- var n int
- var err error
+	var a, b int64
+	var operator string
+	var n int
+	var err error
 
- n, err = fmt.Scanln(&a)
- if err != nil || n != 1 {
-  fmt.Println("Invalid first operand")
-  return 0, 0, ""
- }
+	n, err = fmt.Scanln(&a)
+	if err != nil || n != 1 {
+		fmt.Println("Invalid first operand")
+		return 0, 0, ""
+	}
 
- n, err = fmt.Scanln(&b)
- if err != nil || n != 1 {
-  fmt.Println("Invalid second operand")
-  return 0, 0, ""
- }
+	n, err = fmt.Scanln(&b)
+	if err != nil || n != 1 {
+		fmt.Println("Invalid second operand")
+		return 0, 0, ""
+	}
 
- n, err = fmt.Scanln(&operator)
- if err != nil || n != 1 {
-  fmt.Println("Invalid operation")
-  return 0, 0, ""
- }
+	n, err = fmt.Scanln(&operator)
+	if err != nil || n != 1 {
+		fmt.Println("Invalid operation")
+		return 0, 0, ""
+	}
 
- if operator != "+" && operator != "-" && operator != "*" && operator != "/" {
-  fmt.Println("Invalid operation")
-  return 0, 0, ""
- }
+	if operator != "+" && operator != "-" && operator != "*" && operator != "/" {
+		fmt.Println("Invalid operation")
+		return 0, 0, ""
+	}
 
- return a, b, operator
+	return a, b, operator
 }
